Fixes issue with file v5.14 on Linux
When trying to get the mime type for a file you get the following
error:

> file: invalid option -- 'r'

`-r` is a perfectly valid option, but for some reason it's not
parsing the command properly, whereas the new form is functionally
the same and does work as expected.

I've tracked the issue to the unix `file` command, specifically
version 5.14 - or at least some versions of 5.14 as I've had Ubuntu
boxes both apparently using the same version of the command and one
failing while the other working. I updated the apt package for the
working one and managed to break it without apparently changing the
version!

This is ultimately a workaround for that issue, but there's no
alternative without file being updated as we're reliant on that
command to actually check the MIME type rather than just guessing
from the extension.